requirements.txt
aiogram==3.13.1
asyncpg==0.29.0
timezonefinder==6.5.2
python-dotenv==1.0.1
pytz==2024.2
